+++ src/utils/toast/CustomToastContainer.tsx
@@ -1,0 +1,23 @@
+import { ToastContainer } from "react-toastify";
+import styled from "styled-components";
+
+// TODO : 디자인에 맞게 커스터마이징
+export const CustomToastContainer = styled(ToastContainer)`
+  .Toastify__toast {
+  }
+
+  .Toastify__toast-icon {
+  }
+
+  .Toastify__toast--info {
+  }
+
+  .Toastify__toast--success {
+  }
+
+  .Toastify__toast--error {
+  }
+
+  .Toastify__toast--warning {
+  }
+`;
